+++ class/rexelapi.class.php
@@ -12,7 +12,7 @@
  */
 class RexelApi
 {
-	const CLIENT_VERSION = '1.0.5';
+	const CLIENT_VERSION = '1.0.0';
 	const PRICE_PATH = '/external/productprices/productSalePrices';
 	const STOCK_PATH = '/external/stocks/positions';